Split range_cache.rs into a range_cache/ module and decompose fetch_blocks
- git mv range_cache.rs into range_cache/mod.rs to preserve history, then
  split into error.rs (RangeError, StreamRangesCaller), scheduler.rs
  (Priority, single-flight InFlight/FetchScheduler, FulfillGuard, RunPermit,
  permit-acquire and error-reconstruction helpers), and fetch.rs (the
  RangeCache impl block holding fetch_blocks and its private helpers).
- Decompose fetch_blocks (~294 lines) into four phase helpers on fetch.rs:
  probe_blocks (backend probe + length-mismatch healing + hits/missing
  split), register_missing (single-flight registration, building BatchState
  via a new ::new constructor), spawn_run_fetch (the per-run detached-task
  body, with the success-path chunk splitting further extracted into
  fulfill_run_success), and join_demand/join_prefetch (the two phase-4 join
  strategies).
- Pure refactor: no public API, metric, log, or behavior changes. All moved
  items are pub(super)/private as needed; mod.rs keeps the RangeCache
  struct, its public methods, and the pub re-exports of RangeError and
  StreamRangesCaller.
