--- rust/object-cache/src/range_cache.rs
@@ -1,1290 +1,0 @@
-use std::collections::{BTreeSet, HashMap};
-use std::ops::Range;
-use std::sync::atomic::{AtomicU8, Ordering};
-use std::sync::{Arc, Mutex as StdMutex, Weak};
-use std::time::Instant;
-
-use anyhow::{Result, anyhow};
-use async_stream::try_stream;
-use bytes::{BufMut, Bytes, BytesMut};
-use futures::future::join_all;
-use futures::stream::{self, FuturesUnordered, Stream, StreamExt};
-use micromegas_tracing::prelude::*;
-use object_store::{ObjectStore, ObjectStoreExt, path::Path};
-use tokio::sync::{Notify, OwnedSemaphorePermit, Semaphore, watch};
-
-use super::backend::{BackendDiskStats, FillHint, RangeCacheBackend};
-use super::blocks::{assemble_range, block_byte_range, blocks_for_range, coalesce_runs};
-use super::metric_tags::{self, PrefixTags};
-
-/// Errors returned by [`RangeCache`] that callers may want to handle distinctly.
-#[derive(Debug, thiserror::Error)]
-pub enum RangeError {
-    /// The requested range extends past the end of the object.
-    #[error("requested range end {requested_end} exceeds object size {file_size}")]
-    OutOfBounds { requested_end: u64, file_size: u64 },
-}
-
-/// Which caller is invoking `RangeCache::stream_ranges`, selecting which of
-/// the two distinct `range_cache_get_range_error` / `range_cache_get_ranges_error`
-/// counters is emitted on an upfront validation failure or a mid-stream fetch
-/// error. `stream_ranges` is the single fill path behind both `get_range`/
-/// `get_range_handler` (`Range`) and `get_ranges`/`post_ranges_handler`
-/// (`Ranges`), so the tag keeps the two metric names emitting exactly as they
-/// did before those call sites shared one implementation.
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum StreamRangesCaller {
-    Range,
-    Ranges,
-}
-
-impl StreamRangesCaller {
-    fn emit_error_metric(self) {
-        match self {
-            StreamRangesCaller::Range => imetric!("range_cache_get_range_error", "count", 1_u64),
-            StreamRangesCaller::Ranges => {
-                imetric!("range_cache_get_ranges_error", "count", 1_u64)
-            }
-        }
-    }
-}
-
-pub const DEFAULT_BLOCK_SIZE: u64 = 1024 * 1024;
-
-/// Number of blocks per streamed fetch window used by `stream_ranges`. At the
-/// default 1 MiB block size, 8 blocks (8 MiB) matches one coalesced origin
-/// GET run (`DEFAULT_MAX_COALESCED_GET_BYTES`), with `buffered(2)` giving
-/// modest pipeline overlap. This bounds peak in-flight memory per stream to
-/// roughly `2 * DEMAND_WINDOW_BLOCKS * block_size`, independent of how large
-/// the requested range is.
-pub const DEMAND_WINDOW_BLOCKS: u64 = 8;
-
-/// Default total number of origin GETs allowed to run concurrently. See
-/// `RangeCache::new`.
-pub const DEFAULT_TOTAL_FETCH_PERMITS: usize = 32;
-/// Default number of `DEFAULT_TOTAL_FETCH_PERMITS` slots reserved for demand
-/// reads (never consumed by prefetch).
-pub const DEFAULT_DEMAND_RESERVED_FETCH_PERMITS: usize = 8;
-/// Default max byte span of one coalesced run GET.
-pub const DEFAULT_MAX_COALESCED_GET_BYTES: u64 = 8 * 1024 * 1024;
-/// Default promotion granularity: promote only the run(s) covering a demanded
-/// block, not the whole prefetch batch.
-pub const DEFAULT_PROMOTE_WHOLE_BATCH: bool = false;
-
-/// Concurrency for probing the cache backend for block hits before going to
-/// origin. Backend probes do real disk I/O on a foyer RAM-tier miss, so a
-/// large read probed sequentially would serialize hundreds of disk reads.
-const BACKEND_PROBE_CONCURRENCY: usize = 16;
-
-/// Relative urgency of an origin fetch. Lower is more urgent.
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-#[repr(u8)]
-enum Priority {
-    Demand = 0,
-    Prefetch = 1,
-}
-
-impl Priority {
-    fn from_u8(v: u8) -> Self {
-        if v == Priority::Demand as u8 {
-            Priority::Demand
-        } else {
-            Priority::Prefetch
-        }
-    }
-
-    /// The `class` metric-tag value for this priority (see `metric_tags`).
-    fn class_label(self) -> &'static str {
-        match self {
-            Priority::Demand => metric_tags::CLASS_DEMAND,
-            Priority::Prefetch => metric_tags::CLASS_PREFETCH,
-        }
-    }
-}
-
-/// The `class` tag for one coalesced run: `Demand` if any of its entries is
-/// currently demand priority, else `Prefetch`. Mirrors `acquire_run_permit`'s
-/// own `all_prefetch` check, evaluated once the permit has been acquired so a
-/// promotion that raced the wait is reflected in the tag.
-fn effective_priority(entries: &[Arc<InFlight>]) -> Priority {
-    if entries.iter().any(|e| e.priority() == Priority::Demand) {
-        Priority::Demand
-    } else {
-        Priority::Prefetch
-    }
-}
-
-/// The set of entries submitted in one prefetch call. Used only when
-/// `promote_whole_batch` is enabled: a demand joiner into any sibling entry
-/// promotes every other entry in the batch, not just the one it joined.
-///
-/// Siblings are tracked by `Weak<InFlight>`, captured once each entry is
-/// created or joined, rather than by key string: a key can be removed from
-/// `FetchScheduler::inflight` and later reused by an unrelated fetch (e.g.
-/// after eviction), and promoting by key alone would then spuriously promote
-/// that new, logically unrelated `InFlight`.
-struct BatchState {
-    entries: StdMutex<Vec<Weak<InFlight>>>,
-}
-
-type FetchResult = Result<Bytes, Arc<anyhow::Error>>;
-
-/// One outstanding origin fetch (a single block or a `size()` head), shared
-/// across every concurrent caller asking for the same key.
-struct InFlight {
-    priority: AtomicU8,
-    promote: Notify,
-    result: watch::Sender<Option<FetchResult>>,
-    batch: Option<Arc<BatchState>>,
-    /// Set by the first `fulfill()` call. Lets a `FulfillGuard`'s panic-path
-    /// fallback tell which entries in a partially-completed run already got
-    /// a real result, so it never clobbers one with a synthesized error.
-    fulfilled: std::sync::atomic::AtomicBool,
-}
-
-impl InFlight {
-    fn new(priority: Priority, batch: Option<Arc<BatchState>>) -> Self {
-        let (result, _rx) = watch::channel(None);
-        Self {
-            priority: AtomicU8::new(priority as u8),
-            promote: Notify::new(),
-            result,
-            batch,
-            fulfilled: std::sync::atomic::AtomicBool::new(false),
-        }
-    }
-
-    fn priority(&self) -> Priority {
-        Priority::from_u8(self.priority.load(Ordering::Acquire))
-    }
-
-    fn promote_to_demand(&self) {
-        self.priority
-            .store(Priority::Demand as u8, Ordering::Release);
-        self.promote.notify_one();
-    }
-
-    /// Deliver `result` to every joiner. Idempotent: only the first call
-    /// actually sends — later calls (notably a `FulfillGuard`'s panic-path
-    /// fallback racing a fetch that in fact completed normally) are no-ops,
-    /// so a real result is never overwritten by a synthesized error.
-    fn fulfill(&self, result: FetchResult) {
-        if self.fulfilled.swap(true, Ordering::AcqRel) {
-            return;
-        }
-        // A send error just means every waiter already dropped its
-        // receiver, which is harmless.
-        let _ = self.result.send(Some(result));
-    }
-
-    /// Wait for the fetch to complete, returning immediately if it already
-    /// has. Safe against the "subscribe after send" race: a fresh `watch`
-    /// receiver's initial value counts as unseen until `borrow_and_update`
-    /// checks it, so a result sent before this call is still observed here.
-    async fn join(&self) -> FetchResult {
-        let mut rx = self.result.subscribe();
-        loop {
-            if let Some(r) = rx.borrow_and_update().clone() {
-                return r;
-            }
-            if rx.changed().await.is_err() {
-                return Err(Arc::new(anyhow!(
-                    "in-flight fetch entry dropped without a result"
-                )));
-            }
-        }
-    }
-}
-
-enum Ownership {
-    Owner(Arc<InFlight>),
-    Joiner(Arc<InFlight>),
-}
-
-/// Owns the in-flight single-flight map and the priority-aware origin-fetch
-/// budget. Replaces the two moka caches that used to live directly on
-/// `RangeCache`.
-struct FetchScheduler {
-    inflight: StdMutex<HashMap<String, Arc<InFlight>>>,
-    /// Bounds total concurrent origin GETs (blocks + size heads don't count
-    /// against this; only run GETs acquire it).
-    shared_permits: Arc<Semaphore>,
-    /// Total capacity of `shared_permits`, stored alongside it since
-    /// `tokio::sync::Semaphore` has no capacity accessor. Used by
-    /// `fetch_budget_stats` for the saturation sampler.
-    shared_total: usize,
-    /// Prefetch runs must additionally hold one of these; sized to
-    /// `total - demand_reserved` so demand always finds a free shared permit.
-    prefetch_permits: Arc<Semaphore>,
-    /// Total capacity of `prefetch_permits`, for the same reason as
-    /// `shared_total`.
-    prefetch_total: usize,
-    promote_whole_batch: bool,
-}
-
-impl FetchScheduler {
-    fn new(total: usize, demand_reserved: usize, promote_whole_batch: bool) -> Self {
-        assert!(total > 0, "fetch concurrency total must be > 0");
-        // Strictly less: `demand_reserved == total` would leave the prefetch
-        // semaphore with zero permits, hanging every prefetch run forever.
-        assert!(
-            demand_reserved < total,
-            "demand_reserved ({demand_reserved}) must be < total ({total})"
-        );
-        let prefetch_total = total - demand_reserved;
-        Self {
-            inflight: StdMutex::new(HashMap::new()),
-            shared_permits: Arc::new(Semaphore::new(total)),
-            shared_total: total,
-            prefetch_permits: Arc::new(Semaphore::new(prefetch_total)),
-            prefetch_total,
-            promote_whole_batch,
-        }
-    }
-
-    /// `(shared_available, shared_total, prefetch_available, prefetch_total)`
-    /// -- the fetch-permit budget's current occupancy, for the saturation
-    /// sampler (`object-cache-srv/src/saturation_monitor.rs`).
-    fn fetch_budget_stats(&self) -> (usize, usize, usize, usize) {
-        (
-            self.shared_permits.available_permits(),
-            self.shared_total,
-            self.prefetch_permits.available_permits(),
-            self.prefetch_total,
-        )
-    }
-
-    /// Number of keys (blocks or `size()` heads) currently in flight to
-    /// origin.
-    fn inflight_len(&self) -> usize {
-        self.inflight.lock().expect("inflight lock").len()
-    }
-
-    /// Look up `key` in the in-flight map: become the owner if absent, or a
-    /// joiner if present. A demand joiner into a prefetch-priority entry
-    /// promotes it (and, if `promote_whole_batch`, its batch siblings) to
-    /// demand so it competes for reserved capacity instead of sitting behind
-    /// other prefetch work.
-    ///
-    /// The entry is registered in `batch` (owner or joiner alike) while the
-    /// inflight lock is still held, i.e. before any concurrent demand joiner
-    /// can find the entry and run `promote_batch_siblings` over a
-    /// partially-populated list.
-    fn own_or_join(
-        &self,
-        key: String,
-        prio: Priority,
-        batch: Option<Arc<BatchState>>,
-    ) -> Ownership {
-        let mut promote_batch: Option<(Arc<BatchState>, Arc<InFlight>)> = None;
-        let ownership = {
-            let mut map = self.inflight.lock().expect("inflight lock");
-            if let Some(existing) = map.get(&key) {
-                if prio == Priority::Demand && existing.priority() == Priority::Prefetch {
-                    existing.promote_to_demand();
-                    if self.promote_whole_batch
-                        && let Some(bs) = existing.batch.clone()
-                    {
-                        promote_batch = Some((bs, existing.clone()));
-                    }
-                }
-                if let Some(bs) = &batch {
-                    bs.entries
-                        .lock()
-                        .expect("batch lock")
-                        .push(Arc::downgrade(existing));
-                }
-                Ownership::Joiner(existing.clone())
-            } else {
-                let entry = Arc::new(InFlight::new(prio, batch.clone()));
-                if let Some(bs) = &batch {
-                    bs.entries
-                        .lock()
-                        .expect("batch lock")
-                        .push(Arc::downgrade(&entry));
-                }
-                map.insert(key.clone(), entry.clone());
-                Ownership::Owner(entry)
-            }
-        };
-        if let Some((bs, skip)) = promote_batch {
-            self.promote_batch_siblings(&bs, &skip);
-        }
-        ownership
-    }
-
-    /// Promote every sibling in `batch` other than `skip` to demand priority.
-    /// Siblings are resolved through `Weak<InFlight>` references captured at
-    /// batch-membership time, not by re-looking up their key in `inflight`:
-    /// the key may since have been removed and reused by an unrelated fetch,
-    /// and identity (not the key string) is what must match here.
-    fn promote_batch_siblings(&self, batch: &BatchState, skip: &Arc<InFlight>) {
-        let entries = batch.entries.lock().expect("batch lock");
-        for weak in entries.iter() {
-            if let Some(entry) = weak.upgrade()
-                && !Arc::ptr_eq(&entry, skip)
-            {
-                entry.promote_to_demand();
-            }
-        }
-    }
-
-    fn remove_entry(&self, key: &str) {
-        self.inflight.lock().expect("inflight lock").remove(key);
-    }
-}
-
-/// Scope guard held by the task that owns one or more in-flight entries
-/// (either a single `size()` head or the members of one coalesced block
-/// run). If the owning task exits normally, it calls `disarm()` after
-/// fulfilling every entry and removing it from `FetchScheduler::inflight`.
-///
-/// If instead the task panics (e.g. `Bytes::slice` on a short origin read)
-/// tokio catches the unwind at the task boundary and silently drops the
-/// result, but this guard's `Drop` still runs during that unwind: it
-/// fulfills any not-yet-fulfilled entry with an error (the `fulfilled` flag
-/// on `InFlight` makes this a no-op for entries that already got a real
-/// result) and removes every entry from the map. Without this, a panicking
-/// owner would leave `fulfill()` never called, hanging every joiner
-/// (including the owner itself) forever and leaking the entry permanently.
-struct FulfillGuard {
-    scheduler: Arc<FetchScheduler>,
-    entries: Vec<(String, Arc<InFlight>)>,
-    armed: bool,
-}
-
-impl FulfillGuard {
-    fn new(scheduler: Arc<FetchScheduler>, entries: Vec<(String, Arc<InFlight>)>) -> Self {
-        Self {
-            scheduler,
-            entries,
-            armed: true,
-        }
-    }
-
-    /// Call once the normal completion path has fulfilled and removed every
-    /// entry, so the guard's `Drop` becomes a no-op.
-    fn disarm(mut self) {
-        self.armed = false;
-    }
-}
-
-impl Drop for FulfillGuard {
-    fn drop(&mut self) {
-        if !self.armed {
-            return;
-        }
-        warn!(
-            "fetch task exited without completing normally (likely a panic); \
-             fulfilling {} in-flight entries with an error",
-            self.entries.len()
-        );
-        for (key, entry) in &self.entries {
-            entry.fulfill(Err(Arc::new(anyhow!(
-                "fetch task exited without producing a result (panic during fetch)"
-            ))));
-            self.scheduler.remove_entry(key);
-        }
-    }
-}
-
-/// Held for the duration of one origin GET; dropping it frees the slot(s) for
-/// the next waiter. Fields are never read, only held for their `Drop` effect.
-struct RunPermit {
-    _shared: OwnedSemaphorePermit,
-    _prefetch: Option<OwnedSemaphorePermit>,
-}
-
-/// Resolves when any of `entries`' priority has been promoted since this call
-/// started. Constructing the `Notified` future(s) up front (before the caller
-/// re-checks priority) is what makes this race-free: tokio guarantees a
-/// `notify_one()` that happens after a `Notified` is created is never missed,
-/// even if that `Notified` is later dropped without completing.
-async fn any_entry_promoted(entries: &[Arc<InFlight>]) {
-    match entries {
-        [] => std::future::pending::<()>().await,
-        [only] => only.promote.notified().await,
-        many => {
-            let futs: Vec<_> = many
-                .iter()
-                .map(|e| Box::pin(e.promote.notified()))
-                .collect();
-            futures::future::select_all(futs).await;
-        }
-    }
-}
-
-/// Acquire the permit(s) needed to run one coalesced GET covering `entries`,
-/// honoring promotion: the run's effective priority is the most urgent
-/// (minimum) of its entries', re-checked every time a promotion wakes this
-/// loop so a promotion mid-wait drops the prefetch-class requirement.
-async fn acquire_run_permit(scheduler: &FetchScheduler, entries: &[Arc<InFlight>]) -> RunPermit {
-    loop {
-        let all_prefetch = entries.iter().all(|e| e.priority() == Priority::Prefetch);
-        if !all_prefetch {
-            let shared = scheduler
-                .shared_permits
-                .clone()
-                .acquire_owned()
-                .await
-                .expect("shared_permits semaphore is never closed");
-            return RunPermit {
-                _shared: shared,
-                _prefetch: None,
-            };
-        }
-
-        tokio::select! {
-            prefetch = scheduler.prefetch_permits.clone().acquire_owned() => {
-                let prefetch = prefetch.expect("prefetch_permits semaphore is never closed");
-                tokio::select! {
-                    shared = scheduler.shared_permits.clone().acquire_owned() => {
-                        let shared = shared.expect("shared_permits semaphore is never closed");
-                        return RunPermit { _shared: shared, _prefetch: Some(prefetch) };
-                    }
-                    _ = any_entry_promoted(entries) => {
-                        drop(prefetch);
-                        continue;
-                    }
-                }
-            }
-            _ = any_entry_promoted(entries) => {
-                continue;
-            }
-        }
-    }
-}
-
-/// Reconstruct an owned error from a shared in-flight failure. `anyhow::Error`
-/// is not `Clone`, so a joiner cannot move it out of the `Arc`; and it must
-/// not be stringified, or a missing key would stop downcasting to
-/// `object_store::Error::NotFound` and would surface as a 500 instead of a
-/// 404 (see `validation::is_not_found`).
-fn reconstruct_shared_error(shared: &Arc<anyhow::Error>) -> anyhow::Error {
-    if let Some(object_store::Error::NotFound { path, source }) =
-        shared.downcast_ref::<object_store::Error>()
-    {
-        let rebuilt = object_store::Error::NotFound {
-            path: path.clone(),
-            source: Box::new(std::io::Error::new(
-                std::io::ErrorKind::NotFound,
-                source.to_string(),
-            )),
-        };
-        return anyhow::Error::from(rebuilt);
-    }
-    // `{shared:?}` keeps the full context chain (and backtrace, if captured)
-    // so joiners' 500 log lines are as informative as the owner's.
-    anyhow!("{shared:?}")
-}
-
-fn decode_size(data: &Bytes) -> Result<u64> {
-    Ok(u64::from_le_bytes(
-        data[..8].try_into().expect("8-byte size slice"),
-    ))
-}
-
-/// Range-aware read cache over an origin object store.
-///
-/// # Cache invalidation
-///
-/// This cache assumes object keys are **write-once and content-addressed**: a
-/// given key always maps to the same bytes for the lifetime of the object. The
-/// size and block caches therefore carry no TTL, etag, or generation in their
-/// keys and are never invalidated. Overwriting an existing key with different
-/// content would cause stale size/block data to be served indefinitely. This is
-/// safe for micromegas lake objects (blocks, parquet) which are never
-/// overwritten; do not point this cache at a mutable namespace.
-///
-/// # In-flight map and priority
-///
-/// Concurrent fetches of the same block or size are collapsed via an
-/// in-flight map (`FetchScheduler`): the first caller to ask for a key
-/// becomes its owner and issues the origin request (spawned as a detached
-/// task, so a cancelled caller never strands the others waiting on it);
-/// every other concurrent caller joins and observes the same result.
-/// Contiguous missing blocks the owner controls are coalesced into one
-/// `origin.get_range` per run. Every origin GET is either `Demand` or
-/// `Prefetch` priority; a demand caller joining a prefetch-priority fetch
-/// promotes it (see `own_or_join`), so a late demand read is never stuck
-/// behind unrelated prefetch traffic.
-#[derive(Clone)]
-pub struct RangeCache {
-    origin: Arc<dyn ObjectStore>,
-    backend: Arc<dyn RangeCacheBackend>,
-    block_size: u64,
-    ns: String,
-    scheduler: Arc<FetchScheduler>,
-    max_coalesced_get_bytes: u64,
-    /// Configured `prefix` labels (the server's `allowed_prefixes`, leaked to
-    /// `'static` once at startup), in the same order as `prefix_tags`. Empty
-    /// unless `with_prefix_labels` was used, in which case every key
-    /// classifies as `metric_tags::PREFIX_OTHER`.
-    prefix_labels: Arc<[&'static str]>,
-    /// Precomputed `PrefixTags` parallel to `prefix_labels` (`prefix_tags[i]`
-    /// corresponds to `prefix_labels[i]`), so `classify_tags` is an array
-    /// lookup rather than an allocation + intern-lock per call.
-    prefix_tags: Arc<[PrefixTags]>,
-    /// Precomputed tags for a key matching none of `prefix_labels`.
-    other_tags: PrefixTags,
-}
-
-impl RangeCache {
-    #[allow(clippy::too_many_arguments)]
-    pub fn new(
-        origin: Arc<dyn ObjectStore>,
-        backend: Arc<dyn RangeCacheBackend>,
-        block_size: u64,
-        ns: String,
-        total_fetch_permits: usize,
-        demand_reserved_fetch_permits: usize,
-        max_coalesced_get_bytes: u64,
-        promote_whole_batch: bool,
-    ) -> Self {
-        Self {
-            origin,
-            backend,
-            block_size,
-            ns,
-            scheduler: Arc::new(FetchScheduler::new(
-                total_fetch_permits,
-                demand_reserved_fetch_permits,
-                promote_whole_batch,
-            )),
-            max_coalesced_get_bytes,
-            prefix_labels: Arc::from(Vec::new()),
-            prefix_tags: Arc::from(Vec::new()),
-            other_tags: PrefixTags::new(metric_tags::PREFIX_OTHER),
-        }
-    }
-
-    /// Attach a `prefix` classifier for the dimensioned hit-rate/request
-    /// metrics: `labels` are the server's configured `allowed_prefixes`,
-    /// leaked to `'static` once at startup by the caller (bounded,
-    /// low-cardinality, set once). Every request key is then
-    /// longest-prefix-matched against `labels` (see `classify`); a key
-    /// matching none classifies as `metric_tags::PREFIX_OTHER`.
-    ///
-    /// `RangeCache::new` itself leaves this empty (every key classifies as
-    /// `"other"`), so its existing callers/tests compile unmodified; only
-    /// `object_cache_srv.rs` opts in.
-    pub fn with_prefix_labels(mut self, labels: Arc<[&'static str]>) -> Self {
-        let tags: Vec<PrefixTags> = labels.iter().map(|&label| PrefixTags::new(label)).collect();
-        self.prefix_tags = Arc::from(tags);
-        self.prefix_labels = labels;
-        self
-    }
-
-    /// The precomputed tags for the `prefix` `key` falls under, resolved by
-    /// longest-prefix match against `prefix_labels` (`other_tags` on no
-    /// match). Private: callers needing just the label use `classify`;
-    /// hot-path callers inside this module use the tags directly.
-    fn classify_tags(&self, key: &str) -> &PrefixTags {
-        match metric_tags::longest_prefix_match(&self.prefix_labels, key) {
-            Some(i) => &self.prefix_tags[i],
-            None => &self.other_tags,
-        }
-    }
-
-    /// The `prefix` label `key` falls under (e.g. `"blobs"`, `"views"`, per
-    /// the server's configured `allowed_prefixes`), or
-    /// `metric_tags::PREFIX_OTHER` if it matches none. See
-    /// `with_prefix_labels`.
-    pub fn classify(&self, key: &str) -> &'static str {
-        self.classify_tags(key).label
-    }
-
-    /// `(shared_available, shared_total, prefetch_available, prefetch_total)`
-    /// -- the fetch-permit budget's current occupancy, for the saturation
-    /// sampler.
-    pub fn fetch_budget_stats(&self) -> (usize, usize, usize, usize) {
-        self.scheduler.fetch_budget_stats()
-    }
-
-    /// Number of keys (blocks or `size()` heads) currently in flight to
-    /// origin, for the saturation sampler.
-    pub fn inflight_len(&self) -> usize {
-        self.scheduler.inflight_len()
-    }
-
-    /// Backend disk write-path counters (`None` for a backend with no disk
-    /// tier), for the saturation sampler's per-second foyer disk gauges.
-    pub fn backend_disk_stats(&self) -> Option<BackendDiskStats> {
-        self.backend.disk_stats()
-    }
-
-    /// Size in bytes of one cache block. Every distinct block a request
-    /// touches is fetched and held whole, so callers gating memory (e.g. the
-    /// server's cross-request budget) need this to account for amplification
-    /// from small scattered ranges.
-    pub fn block_size(&self) -> u64 {
-        self.block_size
-    }
-
-    #[span_fn]
-    pub async fn size(&self, key: &str) -> Result<u64> {
-        // The cache key carries no etag/version: see the module docs — keys
-        // are assumed write-once and content-addressed, so a cached size is
-        // never invalidated.
-        let meta_key = format!("meta:{}:{key}", self.ns);
-        let prefix_tag = self.classify_tags(key).prefix;
-
-        if let Some(data) = self.backend.get(&meta_key).await
-            && data.len() == 8
-        {
-            imetric!("range_cache_size_backend_hit", "count", prefix_tag, 1_u64);
-            return decode_size(&data);
-        }
-
-        match self
-            .scheduler
-            .own_or_join(meta_key.clone(), Priority::Demand, None)
-        {
-            Ownership::Owner(entry) => {
-                let origin = self.origin.clone();
-                let backend = self.backend.clone();
-                let scheduler = self.scheduler.clone();
-                let key_owned = key.to_string();
-                let meta_key_owned = meta_key.clone();
-                let task_entry = entry.clone();
-                tokio::spawn(async move {
-                    let guard = FulfillGuard::new(
-                        scheduler.clone(),
-                        vec![(meta_key_owned.clone(), task_entry.clone())],
-                    );
-                    imetric!("range_cache_origin_head", "count", prefix_tag, 1_u64);
-                    let head_path = Path::from(key_owned.as_str());
-                    let head_result = instrument_named!(
-                        origin.head(&head_path),
-                        "range_cache_origin_head_latency"
-                    )
-                    .await;
-                    match head_result {
-                        Ok(object_meta) => {
-                            let size = object_meta.size;
-                            debug!("range_cache origin head key={key_owned} size={size}");
-                            let size_bytes = Bytes::from(size.to_le_bytes().to_vec());
-                            backend
-                                .put(meta_key_owned.clone(), size_bytes.clone(), FillHint::Demand)
-                                .await;
-                            task_entry.fulfill(Ok(size_bytes));
-                        }
-                        Err(e) => {
-                            task_entry.fulfill(Err(Arc::new(anyhow::Error::from(e))));
-                        }
-                    }
-                    scheduler.remove_entry(&meta_key_owned);
-                    guard.disarm();
-                });
-                let data = entry
-                    .join()
-                    .await
-                    .map_err(|e| reconstruct_shared_error(&e))?;
-                decode_size(&data)
-            }
-            Ownership::Joiner(entry) => {
-                let data = entry
-                    .join()
-                    .await
-                    .map_err(|e| reconstruct_shared_error(&e))?;
-                decode_size(&data)
-            }
-        }
-    }
-
-    /// Fetch `indices` (block indices within `key`, sized `file_size`) at the
-    /// given priority, returning the bytes for every requested block on the
-    /// `Demand` path. On the `Prefetch` path bytes are written to the backend
-    /// and dropped as each owned run completes; the returned map is always
-    /// empty, which is what keeps the prefetch peak bounded by
-    /// `prefetch_concurrency * max_coalesced_get_bytes` rather than the full
-    /// request size.
-    async fn fetch_blocks(
-        &self,
-        key: &str,
-        file_size: u64,
-        indices: &[u64],
-        prio: Priority,
-    ) -> Result<HashMap<u64, Bytes>> {
-        if indices.is_empty() {
-            return Ok(HashMap::new());
-        }
-
-        // Classify once per call (a longest-prefix string match), not once
-        // per block probe: `prefix_tags` is reused across every probe and
-        // every coalesced run below.
-        let prefix_tags = self.classify_tags(key);
-        let block_tag = prefix_tags.prefix;
-        let (run_demand_tag, run_prefetch_tag) =
-            (prefix_tags.prefix_demand, prefix_tags.prefix_prefetch);
-
-        let mut hits: HashMap<u64, Bytes> = HashMap::new();
-        let mut missing: Vec<u64> = Vec::new();
-        {
-            // Probe the backend with bounded concurrency: a foyer RAM-tier
-            // miss falls through to async disk I/O, and probing a large read
-            // one block at a time would serialize those disk reads. The
-            // futures are collected eagerly (not mapped lazily inside the
-            // stream) so the resulting future stays `Send`-inferable across
-            // `tokio::spawn` (rustc's HRTB limitation with borrowed closures).
-            let probe_futs: Vec<_> = indices
-                .iter()
-                .map(|&idx| {
-                    let block_key = format!("blk:{}:{key}:{idx}", self.ns);
-                    let backend = self.backend.clone();
-                    async move {
-                        imetric!("range_cache_block_request", "count", block_tag, 1_u64);
-                        let probe =
-                            instrument_named!(backend.get(&block_key), "range_cache_backend_read")
-                                .await;
-                        (idx, probe)
-                    }
-                })
-                .collect();
-            let mut probes = stream::iter(probe_futs).buffer_unordered(BACKEND_PROBE_CONCURRENCY);
-            while let Some((idx, probe)) = probes.next().await {
-                match probe {
-                    Some(data) => {
-                        // A hit's length must match the block's true byte span.
-                        // A mismatch means a prior prefetch stored this block
-                        // under an undersized caller-supplied `size` (or the
-                        // origin object changed size): treat it as a miss so
-                        // it gets refetched and overwritten with the correct
-                        // bytes, healing the poisoned entry.
-                        let expected_range = block_byte_range(idx, self.block_size, file_size);
-                        let expected_len = expected_range.end - expected_range.start;
-                        if data.len() as u64 != expected_len {
-                            imetric!("range_cache_block_len_mismatch", "count", 1_u64);
-                            warn!(
-                                "range_cache block length mismatch key={key} idx={idx} \
-                                 expected={expected_len} got={}",
-                                data.len()
-                            );
-                            missing.push(idx);
-                            continue;
-                        }
-                        imetric!("range_cache_block_backend_hit", "count", block_tag, 1_u64);
-                        // A prefetch has nothing to do with an already-cached
-                        // block, so drop the bytes immediately instead of
-                        // accumulating every hit for the whole call.
-                        if prio == Priority::Demand {
-                            hits.insert(idx, data);
-                        }
-                    }
-                    None => missing.push(idx),
-                }
-            }
-        }
-
-        if missing.is_empty() {
-            return Ok(hits);
-        }
-        missing.sort_unstable();
-        missing.dedup();
-
-        // Only prefetch calls carry a batch handle: it's what lets a demand
-        // joiner promote every sibling of a scattered prefetch call, not just
-        // the one block it happened to touch (see `promote_whole_batch`).
-        let batch = if prio == Priority::Prefetch {
-            Some(Arc::new(BatchState {
-                entries: StdMutex::new(Vec::with_capacity(missing.len())),
-            }))
-        } else {
-            None
-        };
-
-        let mut owned: Vec<u64> = Vec::new();
-        let mut entries: HashMap<u64, Arc<InFlight>> = HashMap::with_capacity(missing.len());
-        for &idx in &missing {
-            let block_key = format!("blk:{}:{key}:{idx}", self.ns);
-            match self.scheduler.own_or_join(block_key, prio, batch.clone()) {
-                Ownership::Owner(entry) => {
-                    owned.push(idx);
-                    entries.insert(idx, entry);
-                }
-                Ownership::Joiner(entry) => {
-                    entries.insert(idx, entry);
-                }
-            }
-        }
-        // Every member of this batch (owned or joined) was recorded as a
-        // promotion sibling inside `own_or_join`, under the inflight lock,
-        // so no entry became joinable before the batch list held it.
-
-        let hint = match prio {
-            Priority::Demand => FillHint::Demand,
-            Priority::Prefetch => FillHint::Prefetch,
-        };
-
-        // `owned` is a sorted subsequence of `missing` (itself sorted), so no
-        // extra sort is needed before coalescing.
-        for run in coalesce_runs(&owned, self.block_size, self.max_coalesced_get_bytes) {
-            let run_len = (run.end - run.start) as usize;
-            let run_entries: Vec<Arc<InFlight>> = (run.start..run.end)
-                .map(|idx| entries.get(&idx).expect("owned entry present").clone())
-                .collect();
-            let run_keys: Vec<String> = (run.start..run.end)
-                .map(|idx| format!("blk:{}:{key}:{idx}", self.ns))
-                .collect();
-            let origin = self.origin.clone();
-            let backend = self.backend.clone();
-            let scheduler = self.scheduler.clone();
-            let block_size = self.block_size;
-            let key_owned = key.to_string();
-
-            // Spawned as a detached task so a cancelled caller (e.g. an axum
-            // request future dropped on client disconnect) can never strand
-            // the other owned blocks or any joiners waiting on this run.
-            tokio::spawn(async move {
-                let guard = FulfillGuard::new(
-                    scheduler.clone(),
-                    run_keys
-                        .iter()
-                        .cloned()
-                        .zip(run_entries.iter().cloned())
-                        .collect(),
-                );
-                let permit_wait_start = Instant::now();
-                let permit = instrument_named!(
-                    acquire_run_permit(&scheduler, &run_entries),
-                    "range_cache_fetch_permit_wait"
-                )
-                .await;
-                // The run's effective priority, resolved once the permit is
-                // acquired so a promotion racing the wait is reflected: used
-                // to tag every latency/hit-rate signal this run emits.
-                let class = effective_priority(&run_entries).class_label();
-                let run_class_tag = if class == metric_tags::CLASS_DEMAND {
-                    run_demand_tag
-                } else {
-                    run_prefetch_tag
-                };
-                fmetric!(
-                    "range_cache_fetch_permit_wait_ms",
-                    "ms",
-                    metric_tags::class_tags(class),
-                    permit_wait_start.elapsed().as_secs_f64() * 1000.0
-                );
-
-                let byte_start = run.start * block_size;
-                let byte_end = block_byte_range(run.end - 1, block_size, file_size).end;
-                let path = Path::from(key_owned.as_str());
-                let origin_get_start = Instant::now();
-                let outcome = instrument_named!(
-                    origin.get_range(&path, byte_start..byte_end),
-                    "range_cache_origin_get"
-                )
-                .await;
-                fmetric!(
-                    "range_cache_origin_get_ms",
-                    "ms",
-                    metric_tags::class_tags(class),
-                    origin_get_start.elapsed().as_secs_f64() * 1000.0
-                );
-                drop(permit);
-
-                match outcome {
-                    Ok(data) => {
-                        // The backend-hit path above heals a length mismatch
-                        // (an undersized cached entry, or the origin object
-                        // having changed size) by treating it as a miss and
-                        // refetching. A true origin fetch has nowhere further
-                        // to fall back to, so a short read here — which
-                        // `object_store`'s `GetRange::Bounded` returns without
-                        // error when the object is shorter than requested —
-                        // must surface as an explicit error instead of being
-                        // silently clipped by `assemble_range`, which would
-                        // otherwise under-yield bytes and either corrupt a
-                        // `Content-Length`-declared response or trip
-                        // `frame_ranges_stream`'s under-yield `.expect()`.
-                        let expected_run_bytes = byte_end - byte_start;
-                        if data.len() as u64 != expected_run_bytes {
-                            imetric!("range_cache_origin_run_len_mismatch", "count", 1_u64);
-                            warn!(
-                                "range_cache origin fetch length mismatch key={key_owned} \
-                                 run={run:?} expected={expected_run_bytes} got={}",
-                                data.len()
-                            );
-                            let shared: Arc<anyhow::Error> = Arc::new(anyhow!(
-                                "origin object changed size mid-fetch: key={key_owned} \
-                                 run={run:?} expected {expected_run_bytes} bytes, got {} bytes",
-                                data.len()
-                            ));
-                            for entry in &run_entries {
-                                entry.fulfill(Err(shared.clone()));
-                            }
-                        } else {
-                            imetric!(
-                                "range_cache_origin_block_fetch",
-                                "count",
-                                run_class_tag,
-                                run_len as u64
-                            );
-                            imetric!(
-                                "range_cache_origin_block_bytes",
-                                "bytes",
-                                run_class_tag,
-                                data.len() as u64
-                            );
-                            debug!(
-                                "range_cache origin fetch key={key_owned} run={run:?} bytes={}",
-                                data.len()
-                            );
-                            for i in 0..run_len {
-                                let offset = i as u64 * block_size;
-                                let local_start = offset as usize;
-                                let local_end =
-                                    (offset + block_size).min(data.len() as u64) as usize;
-                                let chunk = data.slice(local_start..local_end);
-                                backend.put(run_keys[i].clone(), chunk.clone(), hint).await;
-                                run_entries[i].fulfill(Ok(chunk));
-                            }
-                        }
-                    }
-                    Err(e) => {
-                        let shared: Arc<anyhow::Error> = Arc::new(anyhow::Error::from(e));
-                        for entry in &run_entries {
-                            entry.fulfill(Err(shared.clone()));
-                        }
-                    }
-                }
-                for k in &run_keys {
-                    scheduler.remove_entry(k);
-                }
-                guard.disarm();
-            });
-        }
-
-        if prio == Priority::Prefetch {
-            // Join entries as each completes (not in index order), dropping
-            // the joined bytes and the `Arc<InFlight>` right away. The watch
-            // channel inside an entry retains the fulfilled `Bytes` — a slice
-            // sharing its whole run's GET buffer — for as long as the entry
-            // is alive, so holding every entry until all runs finished (as
-            // the demand path's `join_all` does) would let the prefetch peak
-            // reach the full request size instead of the documented
-            // `prefetch_concurrency * max_coalesced_get_bytes` bound. Any
-            // demand joiner that registered before completion still gets the
-            // bytes: it holds its own `Arc<InFlight>` and reads the channel
-            // independently of this early drop.
-            let mut joins: FuturesUnordered<_> = entries
-                .into_values()
-                .map(|entry| async move { entry.join().await.map(|_bytes| ()) })
-                .collect();
-            while let Some(r) = joins.next().await {
-                r.map_err(|e| reconstruct_shared_error(&e))?;
-            }
-            return Ok(HashMap::new());
-        }
-
-        let joined = join_all(missing.iter().map(|idx| {
-            let entry = entries
-                .get(idx)
-                .expect("entry present for every missing index")
-                .clone();
-            let idx = *idx;
-            async move { (idx, entry.join().await) }
-        }))
-        .await;
-
-        for (idx, r) in joined {
-            let data = r.map_err(|e| reconstruct_shared_error(&e))?;
-            hits.insert(idx, data);
-        }
-        Ok(hits)
-    }
-
-    /// Stream the bytes for `ranges` (each half-open `[start, end)`) without
-    /// materializing more than a couple of `DEMAND_WINDOW_BLOCKS` windows of
-    /// blocks at a time, regardless of how large the ranges are in total.
-    ///
-    /// Upfront validation — the `size()` lookup and every range's
-    /// out-of-bounds check — runs before the stream is constructed and
-    /// returned, so 404/416-shaped errors surface synchronously to the
-    /// caller with proper status codes; only a failure *after* that point
-    /// (a mid-stream `fetch_blocks` error, e.g. the origin going down) is
-    /// yielded as the stream's terminal `Err` item. A degenerate
-    /// `start >= end` range is validated like any other (its `end` is still
-    /// checked against `file_size`) but yields no bytes.
-    ///
-    /// Yields a flat, ordered sequence of chunks: ranges are processed in
-    /// the order given and each range's bytes are emitted contiguously
-    /// (possibly split into several `Bytes` chunks at window boundaries)
-    /// before the next range's bytes begin. There is no cross-range block
-    /// dedup — a block shared by two ranges is fetched once per range it
-    /// appears in, though a repeat is always a backend hit or an in-flight
-    /// join, never a second origin GET (see `own_or_join`).
-    ///
-    /// `caller` selects which of the two distinct error counters
-    /// (`range_cache_get_range_error` / `range_cache_get_ranges_error`) this
-    /// call emits on validation failure or a mid-stream fetch error, so
-    /// `get_range`/`get_ranges` (and the two HTTP handlers, which call this
-    /// directly) keep emitting the metric they always have.
-    #[span_fn]
-    pub async fn stream_ranges(
-        &self,
-        key: &str,
-        ranges: Vec<Range<u64>>,
-        caller: StreamRangesCaller,
-    ) -> Result<impl Stream<Item = Result<Bytes>> + Send + 'static> {
-        let file_size = match self.size(key).await {
-            Ok(s) => s,
-            Err(e) => {
-                caller.emit_error_metric();
-                return Err(e);
-            }
-        };
-        self.stream_ranges_inner(key, ranges, file_size, caller)
-            .await
-    }
-
-    /// Like `stream_ranges`, but for a caller that already resolved
-    /// `file_size` itself (e.g. `get_range_handler`, which needs it up front
-    /// for range validation): skips the redundant `self.size()` call, so a
-    /// cache hit doesn't fire `range_cache_size_backend_hit` a second time
-    /// per call.
-    #[span_fn]
-    pub async fn stream_ranges_with_size(
-        &self,
-        key: &str,
-        ranges: Vec<Range<u64>>,
-        file_size: u64,
-        caller: StreamRangesCaller,
-    ) -> Result<impl Stream<Item = Result<Bytes>> + Send + 'static> {
-        self.stream_ranges_inner(key, ranges, file_size, caller)
-            .await
-    }
-
-    async fn stream_ranges_inner(
-        &self,
-        key: &str,
-        ranges: Vec<Range<u64>>,
-        file_size: u64,
-        caller: StreamRangesCaller,
-    ) -> Result<impl Stream<Item = Result<Bytes>> + Send + 'static> {
-        for r in &ranges {
-            if r.end > file_size {
-                caller.emit_error_metric();
-                return Err(RangeError::OutOfBounds {
-                    requested_end: r.end,
-                    file_size,
-                }
-                .into());
-            }
-        }
-
-        let cache = self.clone();
-        let key = key.to_string();
-        Ok(try_stream! {
-            for r in ranges {
-                let start = r.start;
-                let end = r.end;
-                if start >= end {
-                    continue;
-                }
-
-                let blk_range = blocks_for_range(start, end, cache.block_size);
-                let window_indices: Vec<Vec<u64>> = (blk_range.start..blk_range.end)
-                    .collect::<Vec<u64>>()
-                    .chunks(DEMAND_WINDOW_BLOCKS as usize)
-                    .map(|w| w.to_vec())
-                    .collect();
-
-                let mut windows = stream::iter(window_indices)
-                    .map(|w| {
-                        let cache = cache.clone();
-                        let key = key.clone();
-                        async move {
-                            let result = cache
-                                .fetch_blocks(&key, file_size, &w, Priority::Demand)
-                                .await;
-                            (w, result)
-                        }
-                    })
-                    .buffered(2);
-
-                while let Some((w, result)) = windows.next().await {
-                    let block_map = result.inspect_err(|_| caller.emit_error_metric())?;
-                    let blocks: Vec<(u64, Bytes)> = w
-                        .iter()
-                        .map(|idx| {
-                            let data = block_map
-                                .get(idx)
-                                .cloned()
-                                .expect("fetch_blocks returns every requested index");
-                            (*idx, data)
-                        })
-                        .collect();
-                    // Clamp to this window's own byte span (intersected with
-                    // the outer requested range), not the whole range's
-                    // bounds: `blocks` only holds this window's data, and
-                    // `assemble_range` pre-sizes its output buffer from
-                    // `req_end - req_start`, so passing the full range's
-                    // bounds on every window would over-allocate by up to the
-                    // entire range's size on each iteration.
-                    let win_start = w[0] * cache.block_size;
-                    let win_end = block_byte_range(
-                        *w.last().expect("window is non-empty"),
-                        cache.block_size,
-                        file_size,
-                    )
-                    .end;
-                    let local_start = start.max(win_start);
-                    let local_end = end.min(win_end);
-                    yield assemble_range(&blocks, cache.block_size, local_start, local_end);
-                }
-            }
-        })
-    }
-
-    #[span_fn]
-    pub async fn get_range(&self, key: &str, range: Range<u64>) -> Result<Bytes> {
-        let mut stream = Box::pin(
-            self.stream_ranges(key, vec![range], StreamRangesCaller::Range)
-                .await?,
-        );
-        let mut buf = BytesMut::new();
-        while let Some(chunk) = stream.next().await {
-            buf.put_slice(&chunk?);
-        }
-        Ok(buf.freeze())
-    }
-
-    /// Like `get_range`, but for a caller that already resolved `file_size`
-    /// itself, skipping the redundant `self.size()` call inside
-    /// `stream_ranges` (see `stream_ranges_with_size`).
-    #[span_fn]
-    pub async fn get_range_with_size(
-        &self,
-        key: &str,
-        file_size: u64,
-        range: Range<u64>,
-    ) -> Result<Bytes> {
-        let mut stream = Box::pin(
-            self.stream_ranges_with_size(key, vec![range], file_size, StreamRangesCaller::Range)
-                .await?,
-        );
-        let mut buf = BytesMut::new();
-        while let Some(chunk) = stream.next().await {
-            buf.put_slice(&chunk?);
-        }
-        Ok(buf.freeze())
-    }
-
-    #[span_fn]
-    pub async fn get_ranges(&self, key: &str, ranges: &[Range<u64>]) -> Result<Vec<Bytes>> {
-        if ranges.is_empty() {
-            return Ok(vec![]);
-        }
-        let owned_ranges: Vec<Range<u64>> = ranges.to_vec();
-        let stream = Box::pin(
-            self.stream_ranges(key, owned_ranges, StreamRangesCaller::Ranges)
-                .await?,
-        );
-        collect_ranges_from_stream(ranges, stream).await
-    }
-
-    /// Like `get_ranges`, but for a caller that already resolved `file_size`
-    /// itself (see `stream_ranges_with_size`).
-    #[span_fn]
-    pub async fn get_ranges_with_size(
-        &self,
-        key: &str,
-        file_size: u64,
-        ranges: &[Range<u64>],
-    ) -> Result<Vec<Bytes>> {
-        if ranges.is_empty() {
-            return Ok(vec![]);
-        }
-        let owned_ranges: Vec<Range<u64>> = ranges.to_vec();
-        let stream = Box::pin(
-            self.stream_ranges_with_size(key, owned_ranges, file_size, StreamRangesCaller::Ranges)
-                .await?,
-        );
-        collect_ranges_from_stream(ranges, stream).await
-    }
-
-    /// Warm the cache for `ranges` at `Prefetch` priority without returning
-    /// any bytes. The HTTP surface for this (endpoint + client method) is
-    /// #1198; this is the priority-carrying core it builds on. Public (rather
-    /// than crate-private) so integration tests under `tests/` — which
-    /// compile as a separate crate — can exercise the promotion behavior
-    /// described in the fetch-rework plan.
-    pub async fn prefetch_ranges(&self, key: &str, ranges: &[Range<u64>]) -> Result<()> {
-        if ranges.is_empty() {
-            return Ok(());
-        }
-        let file_size = self.size(key).await?;
-        let mut all_block_indices = BTreeSet::new();
-        for r in ranges {
-            let start = r.start;
-            let end = r.end;
-            if end > file_size {
-                return Err(RangeError::OutOfBounds {
-                    requested_end: end,
-                    file_size,
-                }
-                .into());
-            }
-            if start < end {
-                let blk = blocks_for_range(start, end, self.block_size);
-                all_block_indices.extend(blk.start..blk.end);
-            }
-        }
-        self.prefetch_blocks(
-            key,
-            file_size,
-            &all_block_indices.into_iter().collect::<Vec<_>>(),
-        )
-        .await
-    }
-
-    /// Warm the cache for the given block indices at `Prefetch` priority.
-    pub async fn prefetch_blocks(&self, key: &str, file_size: u64, indices: &[u64]) -> Result<()> {
-        self.fetch_blocks(key, file_size, indices, Priority::Prefetch)
-            .await?;
-        Ok(())
-    }
-}
-
-/// Reassemble the flat, ordered chunk sequence a `stream_ranges*` stream
-/// yields back into one `Bytes` per requested range, using each range's
-/// known length rather than relying on a chunk boundary lining up with a
-/// range boundary. Shared by `get_ranges` and `get_ranges_with_size`.
-async fn collect_ranges_from_stream(
-    ranges: &[Range<u64>],
-    mut stream: impl Stream<Item = Result<Bytes>> + Unpin,
-) -> Result<Vec<Bytes>> {
-    let mut result = Vec::with_capacity(ranges.len());
-    let mut pending: Option<Bytes> = None;
-    for r in ranges {
-        let start = r.start;
-        let end = r.end;
-        if start >= end {
-            // `stream_ranges` yields nothing for a degenerate range, so
-            // reinsert the empty chunk at its position ourselves.
-            result.push(Bytes::new());
-            continue;
-        }
-        let need = (end - start) as usize;
-        let mut collected = BytesMut::with_capacity(need);
-        while collected.len() < need {
-            let chunk = match pending.take() {
-                Some(c) => c,
-                None => stream
-                    .next()
-                    .await
-                    .expect("stream_ranges under-yielded for a non-degenerate range")?,
-            };
-            let remaining = need - collected.len();
-            if chunk.len() > remaining {
-                collected.put_slice(&chunk[..remaining]);
-                pending = Some(chunk.slice(remaining..));
-            } else {
-                collected.put_slice(&chunk);
-            }
-        }
-        result.push(collected.freeze());
-    }
-    Ok(result)
-}
